* 'staging' of github.com:construcn/constructn-web: (27 commits)
  logical change In code for display and hiding the loader
  In signup page already exit user try to register again and again message displaying only once issue fixed-1776
  login page issue-1775 is fixed
  fix timeline issue
  In landing page loader is added and also added a message if projects is not assigned to user
  fix build issues
  fix fullscreen issue in compare mode
  placeholder, filter indication and scrollbar
  email converted into lower case and send it to backend
  add fullscreen mode
  848 azure not resolved
  font size for project hierearchy changed
  resolved azure issues
  proj hier border done
  resolved issue list margin top issue
  add issue and task download list
  fixed create issue/task padding
  solved replied comments
  fix download issue
  title value textfield
  ...

# Conflicts:
#	components/container/GenericViewer.jsx

+++ components/divami_components/issue_detail/IssueDetail.tsx
@@ -161,6 +161,7 @@ function BasicTabs(props: any) {
     issueUpdate,
     deleteTheAttachment,
     handleFooter,
+    setTaskState,
   } = props;
 
   const [value, setValue] = React.useState(0);
@@ -276,7 +277,7 @@ function BasicTabs(props: any) {
 
   useEffect(() => {
     if (taskState?.TabOne?.id) {
-      getComments(taskState.TabOne.id);
+      getComments(taskState?.TabOne?.id);
     }
   }, [taskState]);
 
@@ -303,7 +304,7 @@ function BasicTabs(props: any) {
     }
   };
 
-  console.log("taskState.TabOne issue", taskState.TabOne);
+  console.log("taskState.TabOne issue", taskState);
 
   const handleSortMenuClose = () => {
     setIsSortMenuOpen(false);
@@ -405,7 +406,9 @@ function BasicTabs(props: any) {
             <div></div>
             <Image
               src={
-                taskState.TabOne.screenshot ? taskState.TabOne.screenshot : ""
+                taskState?.TabOne?.screenshot
+                  ? taskState?.TabOne?.screenshot
+                  : ""
               }
               alt=""
               width={400}
@@ -416,14 +419,14 @@ function BasicTabs(props: any) {
             <SecondContPrior>
               <PriorityTitle>Type</PriorityTitle>
               <PriorityStatus style={{ color: "#101F4B" }}>
-                {taskState.TabOne.type}
+                {taskState?.TabOne?.type}
               </PriorityStatus>
             </SecondContPrior>
 
             <SecondContPriorParal>
               <PriorityTitle>Priority</PriorityTitle>
               <PriorityStatus style={{ color: "#101F4B" }}>
-                {taskState.TabOne.priority}
+                {taskState?.TabOne?.priority}
               </PriorityStatus>
             </SecondContPriorParal>
           </SecondBodyDiv>
@@ -433,7 +436,7 @@ function BasicTabs(props: any) {
               <CaptureTitle>Captured on</CaptureTitle>
               <CaptureStatus style={{ color: "#101F4B" }}>
                 {" "}
-                {Moment(taskState.TabOne.capturedOn).format("DD MMM YYYY")}
+                {Moment(taskState?.TabOne?.capturedOn).format("DD MMM YYYY")}
               </CaptureStatus>
             </SecondContCapt>
 
@@ -441,7 +444,7 @@ function BasicTabs(props: any) {
               <ThirdContWatch>Watcher</ThirdContWatch>
               <ThirdContWatchName style={{ color: "#101F4B" }}>
                 {" "}
-                {taskState.TabOne.creator}
+                {taskState?.TabOne?.creator}
               </ThirdContWatchName>
             </SecondContPriorParal>
           </SecondBodyDiv>
@@ -520,8 +523,8 @@ function BasicTabs(props: any) {
                 <ThirdContProg>Progress</ThirdContProg>
 
                 <ThirdContProgType style={{ color: "#101F4B" }}>
-                  {taskState.TabOne.status}
-                  {taskState.TabOne.status ? (
+                  {taskState?.TabOne?.status}
+                  {taskState?.TabOne?.status ? (
                     <PenIconImage
                       onClick={() => {
                         handleEditProgress();
@@ -611,7 +614,7 @@ function BasicTabs(props: any) {
                 config={progressOptionsState[0]}
                 data={{
                   ...progressOptionsState[0],
-                  defaultValue: taskState.TabOne.status,
+                  defaultValue: taskState?.TabOne?.status,
                 }}
                 // defaultValue={progressOptionsState?.options[0].value}
                 id={"issuePriority"}
@@ -718,7 +721,7 @@ function BasicTabs(props: any) {
               <DescriptionTitle>Issue Description</DescriptionTitle>
 
               <DescriptionPara>
-                {taskState.TabOne.issueDescription}
+                {taskState?.TabOne?.issueDescription}
               </DescriptionPara>
             </DescriptionDiv>
           ) : (
@@ -747,6 +750,19 @@ function BasicTabs(props: any) {
                               alt={"delete icon"}
                               onClick={() => {
                                 deleteTheAttachment(a?._id, "issue");
+                                setTaskState((prev: any) => {
+                                  const updatedTabOne = {
+                                    ...prev.TabOne,
+                                    attachments: prev.TabOne.attachments.filter(
+                                      (attachment: any) =>
+                                        attachment._id !== a?._id
+                                    ),
+                                  };
+                                  return {
+                                    ...prev,
+                                    TabOne: updatedTabOne,
+                                  };
+                                });
                               }}
                               className={`deleteIcon`}
                             />
@@ -824,7 +840,7 @@ function BasicTabs(props: any) {
                   </AttachButton> */}
                       <SendButton
                         onClick={() => {
-                          addComment(comments, taskState.TabOne.id);
+                          addComment(comments, taskState?.TabOne?.id);
                         }}
                       >
                         <ImageErrorIcon src={Send} alt="" />
@@ -862,7 +878,7 @@ function BasicTabs(props: any) {
             </AttachButton>
             <SendButton
               onClick={() => {
-                addComment(comments, taskState.TabOne.id);
+                addComment(comments, taskState?.TabOne?.id);
               }}
             >
               <ImageErrorIcon src={Send} alt="" />
@@ -903,7 +919,7 @@ const CustomIssueDetailsDrawer = (props: any) => {
 
   const onDeleteIssue = (status: any) => {
     setshowPopUp(false);
-    deleteTheIssue(selectedIssue, onClose);
+    if (deleteTheIssue) deleteTheIssue(selectedIssue, onClose);
   };
   // const deleteTheAttachment = (attachmentId: string) => {
   //   deleteAttachment(attachmentId)
@@ -1225,6 +1241,7 @@ const CustomIssueDetailsDrawer = (props: any) => {
             issueUpdate={issueUpdate}
             deleteTheAttachment={deleteTheAttachment}
             handleFooter={SetFooterState}
+            setTaskState={setTaskState}
           />
         </BodyContainer>
       </CustomTaskDrawerContainer>
